+++ Cosette/Engine/Ai/Score/PieceSquareTables/QueenTables.cs
@@ -7,24 +7,24 @@ namespace Cosette.Engine.Ai.Score.PieceSquareTables
         public static int[][] Pattern =
         {
             // Opening
-            new[] { -15, -10, -10,  -5,  -5, -10, -10,  -15,
-                    -10,   0,   0,   0,   0,   0,   0,  -10,
-                    -10,   0,   5,   5,   5,   5,   0,  -10,
-                     -5,   0,   5,  10,  10,   5,   0,   -5,
-                     -5,   0,   5,  10,  10,   5,   0,   -5,
-                    -10,   0,   5,   5,   5,   5,   0,  -10,
-                    -10,   0,   0,   0,   0,   0,   0,  -10,
-                    -15, -10, -10,  -5,  -5,  -10, -10, -15 },
+            new[] { -15, -10, -10,  -5,  -5, -10, -10, -15,
+                    -10,   0,   0,   0,   0,   0,   0, -10,
+                    -10,   0,   5,   5,   5,   5,   0, -10,
+                     -5,   0,   5,  10,  10,   5,   0,  -5,
+                     -5,   0,   5,  10,  10,   5,   0,  -5,
+                    -10,   0,   5,   5,   5,   5,   0, -10,
+                    -10,   0,   0,   0,   0,   0,   0, -10,
+                    -15, -10, -10,  -5,  -5, -10, -10, -15 },
 
             // Ending
-            new[] { 5,   5,   5,   5,   5,   5,   5,   5,
-                    5,   5,   5,   5,   5,   5,   5,   5,
-                    5,   5,   5,   5,   5,   5,   5,   5,
-                    5,   5,   5,   5,   5,   5,   5,   5,
-                    5,   5,   5,   5,   5,   5,   5,   5,
-                    5,   5,   5,   5,   5,   5,   5,   5,
-                    5,   5,   5,   5,   5,   5,   5,   5,
-                    5,   5,   5,   5,   5,   5,   5,   5}
+            new[] {   5,   5,   5,   5,   5,   5,   5,   5,
+                      5,   5,   5,   5,   5,   5,   5,   5,
+                      5,   5,   5,   5,   5,   5,   5,   5,
+                      5,   5,   5,   5,   5,   5,   5,   5,
+                      5,   5,   5,   5,   5,   5,   5,   5,
+                      5,   5,   5,   5,   5,   5,   5,   5,
+                      5,   5,   5,   5,   5,   5,   5,   5,
+                      5,   5,   5,   5,   5,   5,   5,   5 }
         };
 
         public static int[][][] Values =
@@ -32,15 +32,15 @@ namespace Cosette.Engine.Ai.Score.PieceSquareTables
             // White
             new []
             {
-                TableOperations.FlipVertically(Pattern[(int)GamePhase.Opening]),
-                TableOperations.FlipVertically(Pattern[(int)GamePhase.Ending])
+                TableOperations.FlipVertically(Pattern[GamePhase.Opening]),
+                TableOperations.FlipVertically(Pattern[GamePhase.Ending])
             },
 
             // Black
             new []
             {
-                TableOperations.FlipHorizontally(Pattern[(int)GamePhase.Opening]),
-                TableOperations.FlipHorizontally(Pattern[(int)GamePhase.Ending])
+                TableOperations.FlipHorizontally(Pattern[GamePhase.Opening]),
+                TableOperations.FlipHorizontally(Pattern[GamePhase.Ending])
             }
         };
     }
